chore(api): remove dead code and consolidate duplicated tests
- Delete unused calculateLegacyPositionPnL and calculatePositionPnL (no callers outside their own tests); keep calculateCombinedPositionPnL
- Delete graphql/resolvers/GeneratedResolvers.ts (unused startGraphQLServer; prisma re-export replaced with direct import from db)
- Delete test-prediction-indexer-e2e.ts and orphan test:prediction-*, test:websocket, test:rate-limiting package.json scripts (pointed to missing files, not in CI)
- Merge routes/__tests__/conditions.test.ts into routes/conditions.test.ts (duplicate file created in #opt-name-updates; only the optionName batch tests were unique)
- Drop stale graphql-audit._ljm_.md doc reference
