+++ packages/api/package.json
@@ -27,13 +27,6 @@
     "pretest": "pnpm --filter @sapience/sdk run build:lib",
     "test": "vitest run",
     "test:watch": "vitest",
-    "test:prediction-e2e": "tsx test-prediction-indexer-e2e.ts",
-    "test:prediction-mint": "tsx test-prediction-indexer-e2e.ts mint",
-    "test:prediction-burn": "tsx test-prediction-indexer-e2e.ts burn",
-    "test:prediction-both": "tsx test-prediction-indexer-e2e.ts both",
-    "test:prediction-assertion": "tsx test-prediction-indexer-e2e.ts assertion",
-    "test:websocket": "tsx test-websocket-connection.ts",
-    "test:rate-limiting": "tsx src/scripts/testRateLimiting.ts",
     "emit-schema": "tsx src/scripts/emitSchema.ts",
     "generate-types": "pnpm run prisma:generate && pnpm run emit-schema && graphql-codegen --config codegen.ts",
     "sentry:sourcemaps": "sentry-cli sourcemaps inject --org foil --project data ./dist && sentry-cli sourcemaps upload --org foil --project data ./dist",
